Add externalsecret for NERC OpenShift
The environment variable needs to match the name of the resource.
Given that the resource type will be shown in parentheses, the
name of OCP Prod will show up as NERC (OpenShift).

I've also added the same secret a second time for a possible
NERC Experimental resource name, which we can later rename to
just NERC without needing a restart or redeploy.