Allow validation of Openshift allocations when changing quotas
In combination with changes to `coldfront-plugin-cloud` [1],
a Coldfront patch is added to allow validation of
Openshift allocation change requests (arc).

Currently only one scenario is tested:
If the acr requests for quotas below the current usage of
the project, the user will be sent back to the change
request form with an informative error message.

This feature is possible by leveraging Django formsets
and their built-in custom validation feature [2]

[1] https://github.com/nerc-project/coldfront-plugin-cloud/pull/208
[2] https://docs.djangoproject.com/en/5.1/topics/forms/formsets/#custom-formset-validation